changed Search to functional component and moved google maps functions from CreateEvent
